requirements.txt
requests==2.21.0
psycopg2==2.8.5
requests_oauthlib==0.5.0
Werkzeug==0.16.0
celery==4.4.6
Flask==1.1.2
python-dotenv==0.14.0
bcrypt
